Move gql in group
