+++ src/ai/backend/manager/models/group.py
@@ -20,26 +20,18 @@ from typing import (
     override,
 )
 
-import graphene
 import sqlalchemy as sa
 import trafaret as t
-from graphene.types.datetime import DateTime as GQLDateTime
-from graphql import Undefined
-from sqlalchemy.engine.row import Row
 from sqlalchemy.ext.asyncio import AsyncConnection as SAConnection
 from sqlalchemy.ext.asyncio import AsyncSession
 from sqlalchemy.orm import load_only, relationship, selectinload
 from sqlalchemy.orm.exc import NoResultFound
 
 from ai.backend.common import msgpack
-from ai.backend.common.types import ResourceSlot
 from ai.backend.logging import BraceStyleAdapter
 from ai.backend.manager.models.association_container_registries_groups import (
     AssociationContainerRegistriesGroupsRow,
 )
-from ai.backend.manager.models.utils import define_state
-from ai.backend.manager.services.groups.types import GroupCreator, GroupModifier
-from ai.backend.manager.types import OptionalState
 
 from ..defs import RESERVED_DOTFILES
 from .base import (
@@ -51,10 +43,7 @@ from .base import (
     SlugType,
     StructuredJSONColumn,
     VFolderHostPermissionColumn,
-    batch_multiresult,
-    batch_result,
     mapper_registry,
-    privileged_mutation,
 )
 from .rbac import (
     AbstractPermissionContext,
@@ -69,30 +58,9 @@ from .rbac import (
 )
 from .rbac.context import ClientContext
 from .rbac.permission_defs import ProjectPermission
-from .user import UserRole
 
 if TYPE_CHECKING:
-    from ai.backend.manager.services.groups.actions.create_group import (
-        CreateGroupAction,
-        CreateGroupActionResult,
-    )
-    from ai.backend.manager.services.groups.actions.delete_group import (
-        DeleteGroupAction,
-        DeleteGroupActionResult,
-    )
-    from ai.backend.manager.services.groups.actions.modify_group import (
-        ModifyGroupAction,
-        ModifyGroupActionResult,
-    )
-    from ai.backend.manager.services.groups.actions.purge_group import (
-        PurgeGroupAction,
-        PurgeGroupActionResult,
-    )
-    from ai.backend.manager.services.groups.types import GroupData
-
-    from .gql import GraphQueryContext
     from .rbac import ContainerRegistryScope
-    from .scaling_group import ScalingGroup
 
 log = BraceStyleAdapter(logging.getLogger(__spec__.name))
 
@@ -103,12 +71,6 @@ __all__: Sequence[str] = (
     "association_groups_users",
     "AssocGroupUserRow",
     "resolve_group_name_or_id",
-    "Group",
-    "GroupInput",
-    "ModifyGroupInput",
-    "CreateGroup",
-    "ModifyGroup",
-    "DeleteGroup",
     "GroupDotfile",
     "ProjectType",
     "MAXIMUM_DOTFILE_SIZE",
@@ -392,466 +354,6 @@ async def resolve_groups(
     return_val = [row["id"] for row in rows]
 
     return return_val
-
-
-class Group(graphene.ObjectType):
-    id = graphene.UUID()
-    name = graphene.String()
-    description = graphene.String()
-    is_active = graphene.Boolean()
-    created_at = GQLDateTime()
-    modified_at = GQLDateTime()
-    domain_name = graphene.String()
-    total_resource_slots = graphene.JSONString()
-    allowed_vfolder_hosts = graphene.JSONString()
-    integration_id = graphene.String()
-    resource_policy = graphene.String()
-    type = graphene.String(description="Added in 24.03.0.")
-    container_registry = graphene.JSONString(description="Added in 24.03.0.")
-
-    scaling_groups = graphene.List(lambda: graphene.String)
-
-    @classmethod
-    def from_row(cls, graph_ctx: GraphQueryContext, row: Row) -> Optional[Group]:
-        if row is None:
-            return None
-        return cls(
-            id=row["id"],
-            name=row["name"],
-            description=row["description"],
-            is_active=row["is_active"],
-            created_at=row["created_at"],
-            modified_at=row["modified_at"],
-            domain_name=row["domain_name"],
-            total_resource_slots=(
-                row["total_resource_slots"].to_json()
-                if row["total_resource_slots"] is not None
-                else {}
-            ),
-            allowed_vfolder_hosts=row["allowed_vfolder_hosts"].to_json(),
-            integration_id=row["integration_id"],
-            resource_policy=row["resource_policy"],
-            type=row["type"].name,
-            container_registry=row["container_registry"],
-        )
-
-    @classmethod
-    def from_dto(cls, dto: Optional[GroupData]) -> Optional[Self]:
-        if dto is None:
-            return None
-        return cls(
-            id=dto.id,
-            name=dto.name,
-            description=dto.description,
-            is_active=dto.is_active,
-            created_at=dto.created_at,
-            modified_at=dto.modified_at,
-            domain_name=dto.domain_name,
-            total_resource_slots=dto.total_resource_slots.to_json()
-            if dto.total_resource_slots
-            else {},
-            allowed_vfolder_hosts=dto.allowed_vfolder_hosts.to_json(),
-            integration_id=dto.integration_id,
-            resource_policy=dto.resource_policy,
-            type=dto.type.name,
-            container_registry=dto.container_registry,
-        )
-
-    async def resolve_scaling_groups(self, info: graphene.ResolveInfo) -> Sequence[ScalingGroup]:
-        graph_ctx: GraphQueryContext = info.context
-        loader = graph_ctx.dataloader_manager.get_loader(
-            graph_ctx,
-            "ScalingGroup.by_group",
-        )
-        sgroups = await loader.load(self.id)
-        return [sg.name for sg in sgroups]
-
-    @classmethod
-    async def load_all(
-        cls,
-        graph_ctx: GraphQueryContext,
-        *,
-        domain_name: Optional[str] = None,
-        is_active: Optional[bool] = None,
-        type: list[ProjectType] = [ProjectType.GENERAL],
-    ) -> Sequence[Group]:
-        query = sa.select([groups]).select_from(groups).where(groups.c.type.in_(type))
-        if domain_name is not None:
-            query = query.where(groups.c.domain_name == domain_name)
-        if is_active is not None:
-            query = query.where(groups.c.is_active == is_active)
-        async with graph_ctx.db.begin_readonly() as conn:
-            return [
-                obj
-                async for row in (await conn.stream(query))
-                if (obj := cls.from_row(graph_ctx, row)) is not None
-            ]
-
-    @classmethod
-    async def batch_load_by_id(
-        cls,
-        graph_ctx: GraphQueryContext,
-        group_ids: Sequence[uuid.UUID],
-        *,
-        domain_name: Optional[str] = None,
-    ) -> Sequence[Group | None]:
-        query = sa.select([groups]).select_from(groups).where(groups.c.id.in_(group_ids))
-        if domain_name is not None:
-            query = query.where(groups.c.domain_name == domain_name)
-        async with graph_ctx.db.begin_readonly() as conn:
-            return await batch_result(
-                graph_ctx,
-                conn,
-                query,
-                cls,
-                group_ids,
-                lambda row: row["id"],
-            )
-
-    @classmethod
-    async def batch_load_by_name(
-        cls,
-        graph_ctx: GraphQueryContext,
-        group_names: Sequence[str],
-        *,
-        domain_name: Optional[str] = None,
-    ) -> Sequence[Sequence[Group | None]]:
-        query = sa.select([groups]).select_from(groups).where(groups.c.name.in_(group_names))
-        if domain_name is not None:
-            query = query.where(groups.c.domain_name == domain_name)
-        async with graph_ctx.db.begin_readonly() as conn:
-            return await batch_multiresult(
-                graph_ctx,
-                conn,
-                query,
-                cls,
-                group_names,
-                lambda row: row["name"],
-            )
-
-    @classmethod
-    async def batch_load_by_user(
-        cls,
-        graph_ctx: GraphQueryContext,
-        user_ids: Sequence[uuid.UUID],
-        *,
-        type: list[ProjectType] | None = None,
-    ) -> Sequence[Sequence[Group | None]]:
-        if type is None:
-            _type = [ProjectType.GENERAL]
-        else:
-            _type = type
-        j = sa.join(
-            groups,
-            association_groups_users,
-            groups.c.id == association_groups_users.c.group_id,
-        )
-        query = (
-            sa.select([groups, association_groups_users.c.user_id])
-            .select_from(j)
-            .where(association_groups_users.c.user_id.in_(user_ids) & (groups.c.type.in_(_type)))
-        )
-        async with graph_ctx.db.begin_readonly() as conn:
-            return await batch_multiresult(
-                graph_ctx,
-                conn,
-                query,
-                cls,
-                user_ids,
-                lambda row: row["user_id"],
-            )
-
-    @classmethod
-    async def get_groups_for_user(
-        cls,
-        graph_ctx: GraphQueryContext,
-        user_id: uuid.UUID,
-    ) -> Sequence[Group]:
-        j = sa.join(
-            groups,
-            association_groups_users,
-            groups.c.id == association_groups_users.c.group_id,
-        )
-        query = (
-            sa.select([groups]).select_from(j).where(association_groups_users.c.user_id == user_id)
-        )
-        async with graph_ctx.db.begin_readonly() as conn:
-            return [
-                obj
-                async for row in (await conn.stream(query))
-                if (obj := cls.from_row(graph_ctx, row)) is not None
-            ]
-
-
-class GroupInput(graphene.InputObjectType):
-    type = graphene.String(
-        required=False,
-        default_value="GENERAL",
-        description=(
-            f"Added in 24.03.0. Available values: {', '.join([p.name for p in ProjectType])}"
-        ),
-    )
-    description = graphene.String(required=False, default_value="")
-    is_active = graphene.Boolean(required=False, default_value=True)
-    domain_name = graphene.String(required=True)
-    total_resource_slots = graphene.JSONString(required=False, default_value={})
-    allowed_vfolder_hosts = graphene.JSONString(required=False, default_value={})
-    integration_id = graphene.String(required=False, default_value="")
-    resource_policy = graphene.String(required=False, default_value="default")
-    container_registry = graphene.JSONString(
-        required=False, default_value={}, description="Added in 24.03.0"
-    )
-
-    def to_action(self, name: str) -> CreateGroupAction:
-        def value_or_none(value):
-            return value if value is not Undefined else None
-
-        type_val = None if self.type is Undefined else ProjectType[self.type]
-        description_val = value_or_none(self.description)
-        is_active_val = value_or_none(self.is_active)
-        total_resource_slots_val = (
-            None
-            if self.total_resource_slots is Undefined
-            else ResourceSlot.from_user_input(self.total_resource_slots, None)
-        )
-        allowed_vfolder_hosts_val = value_or_none(self.allowed_vfolder_hosts)
-        integration_id_val = value_or_none(self.integration_id)
-        resource_policy_val = value_or_none(self.resource_policy)
-        container_registry_val = value_or_none(self.container_registry)
-
-        return CreateGroupAction(
-            input=GroupCreator(
-                name=name,
-                domain_name=self.domain_name,
-                type=type_val,
-                description=description_val,
-                is_active=is_active_val,
-                total_resource_slots=total_resource_slots_val,
-                allowed_vfolder_hosts=allowed_vfolder_hosts_val,
-                integration_id=integration_id_val,
-                resource_policy=resource_policy_val,
-                container_registry=container_registry_val,
-            ),
-        )
-
-
-class ModifyGroupInput(graphene.InputObjectType):
-    name = graphene.String(required=False)
-    description = graphene.String(required=False)
-    is_active = graphene.Boolean(required=False)
-    domain_name = graphene.String(required=False)
-    total_resource_slots = graphene.JSONString(required=False)
-    user_update_mode = graphene.String(required=False)
-    user_uuids = graphene.List(lambda: graphene.String, required=False)
-    allowed_vfolder_hosts = graphene.JSONString(required=False)
-    integration_id = graphene.String(required=False)
-    resource_policy = graphene.String(required=False)
-    container_registry = graphene.JSONString(
-        required=False, default_value={}, description="Added in 24.03.0"
-    )
-
-    def to_action(self, group_id: uuid.UUID) -> ModifyGroupAction:
-        def value_or_none(value):
-            return value if value is not Undefined else None
-
-        return ModifyGroupAction(
-            group_id=group_id,
-            modifier=GroupModifier(
-                name=OptionalState(
-                    "name",
-                    define_state(self.name),
-                    value_or_none(self.name),
-                ),
-                domain_name=OptionalState(
-                    "domain_name",
-                    define_state(self.domain_name),
-                    value_or_none(self.domain_name),
-                ),
-                description=OptionalState(
-                    "description",
-                    define_state(self.description),
-                    value_or_none(self.description),
-                ),
-                is_active=OptionalState(
-                    "is_active",
-                    define_state(self.is_active),
-                    value_or_none(self.is_active),
-                ),
-                total_resource_slots=OptionalState(
-                    "total_resource_slots",
-                    define_state(self.total_resource_slots),
-                    None
-                    if self.total_resource_slots is Undefined
-                    else ResourceSlot.from_user_input(self.total_resource_slots, None),
-                ),
-                user_update_mode=OptionalState(
-                    "user_update_mode",
-                    define_state(self.user_update_mode),
-                    value_or_none(self.user_update_mode),
-                ),
-                user_uuids=OptionalState(
-                    "user_uuids",
-                    define_state(self.user_uuids),
-                    value_or_none(self.user_uuids),
-                ),
-                allowed_vfolder_hosts=OptionalState(
-                    "allowed_vfolder_hosts",
-                    define_state(self.allowed_vfolder_hosts),
-                    value_or_none(self.allowed_vfolder_hosts),
-                ),
-                integration_id=OptionalState(
-                    "integration_id",
-                    define_state(self.integration_id),
-                    value_or_none(self.integration_id),
-                ),
-                resource_policy=OptionalState(
-                    "resource_policy",
-                    define_state(self.resource_policy),
-                    value_or_none(self.resource_policy),
-                ),
-                container_registry=OptionalState(
-                    "container_registry",
-                    define_state(self.container_registry),
-                    value_or_none(self.container_registry),
-                ),
-            ),
-        )
-
-
-class CreateGroup(graphene.Mutation):
-    allowed_roles = (UserRole.ADMIN, UserRole.SUPERADMIN)
-
-    class Arguments:
-        name = graphene.String(required=True)
-        props = GroupInput(required=True)
-
-    ok = graphene.Boolean()
-    msg = graphene.String()
-    group = graphene.Field(lambda: Group, required=False)
-
-    @classmethod
-    @privileged_mutation(
-        UserRole.ADMIN,
-        lambda name, props, **kwargs: (props.domain_name, None),
-    )
-    async def mutate(
-        cls,
-        root,
-        info: graphene.ResolveInfo,
-        name: str,
-        props: GroupInput,
-    ) -> CreateGroup:
-        graph_ctx: GraphQueryContext = info.context
-
-        action = props.to_action(name)
-        res: CreateGroupActionResult = (
-            await graph_ctx.processors.group.create_group.wait_for_complete(action)
-        )
-
-        return cls(
-            ok=res.success,
-            msg="success" if res.success else "failed",
-            group=Group.from_dto(res.data) if res.success else None,
-        )
-
-
-class ModifyGroup(graphene.Mutation):
-    allowed_roles = (UserRole.ADMIN, UserRole.SUPERADMIN)
-
-    class Arguments:
-        gid = graphene.UUID(required=True)
-        props = ModifyGroupInput(required=True)
-
-    ok = graphene.Boolean()
-    msg = graphene.String()
-    group = graphene.Field(lambda: Group, required=False)
-
-    @classmethod
-    @privileged_mutation(
-        UserRole.ADMIN,
-        lambda gid, **kwargs: (None, gid),
-    )
-    async def mutate(
-        cls,
-        root,
-        info: graphene.ResolveInfo,
-        gid: uuid.UUID,
-        props: ModifyGroupInput,
-    ) -> ModifyGroup:
-        graph_ctx: GraphQueryContext = info.context
-
-        action = props.to_action(gid)
-        res: ModifyGroupActionResult = (
-            await graph_ctx.processors.group.modify_group.wait_for_complete(action)
-        )
-
-        return cls(
-            ok=res.success,
-            msg="success" if res.success else "failed",
-            group=Group.from_dto(res.data) if res.success else None,
-        )
-
-
-class DeleteGroup(graphene.Mutation):
-    """
-    Instead of deleting the group, just mark it as inactive.
-    """
-
-    allowed_roles = (UserRole.ADMIN, UserRole.SUPERADMIN)
-
-    class Arguments:
-        gid = graphene.UUID(required=True)
-
-    ok = graphene.Boolean()
-    msg = graphene.String()
-
-    @classmethod
-    @privileged_mutation(
-        UserRole.ADMIN,
-        lambda gid, **kwargs: (None, gid),
-    )
-    async def mutate(cls, root, info: graphene.ResolveInfo, gid: uuid.UUID) -> DeleteGroup:
-        ctx: GraphQueryContext = info.context
-        res: DeleteGroupActionResult = await ctx.processors.group.delete_group.wait_for_complete(
-            DeleteGroupAction(gid)
-        )
-
-        return cls(ok=res.success, msg="success" if res.success else "failed")
-
-
-class PurgeGroup(graphene.Mutation):
-    """
-    Completely deletes a group from DB.
-
-    Group's vfolders and their data will also be lost
-    as well as the kernels run from the group.
-    There is no migration of the ownership for group folders.
-    """
-
-    allowed_roles = (UserRole.ADMIN, UserRole.SUPERADMIN)
-
-    class Arguments:
-        gid = graphene.UUID(required=True)
-
-    ok = graphene.Boolean()
-    msg = graphene.String()
-
-    @classmethod
-    @privileged_mutation(
-        UserRole.ADMIN,
-        lambda gid, **kwargs: (None, gid),
-    )
-    async def mutate(cls, root, info: graphene.ResolveInfo, gid: uuid.UUID) -> PurgeGroup:
-        graph_ctx: GraphQueryContext = info.context
-
-        res: PurgeGroupActionResult = (
-            await graph_ctx.processors.group.purge_group.wait_for_complete(PurgeGroupAction(gid))
-        )
-
-        return cls(
-            ok=res.success,
-            msg="success" if res.success else "failed",
-        )
 
 
 class GroupDotfile(TypedDict):
